Update research.html
Added reference to SECVULEVAL paper which was the catalyst for the start of the project. I have several draft research MD documents showing how the 6D methodology works to fill in the gaps discussed in the paper. The papers are not in the repo; they are pending my review. But I am posting the basic finding as an impact to the overall project. I did not post the study earlier as I wanted to  test some of my assumptions first.